+++ src/main/java/net/lopymine/ip/element/base/IParticle.java
@@ -1,6 +1,4 @@
-package net.lopymine.ip.renderer;
-
-import net.lopymine.ip.element.base.*;
+package net.lopymine.ip.element.base;
 
 public interface IParticle extends IAlive, IRenderable, ISelectableElement, IHoverableElement, IMovableElement {
 }