+++ src/main/java/net/lopymine/ip/entrypoint/IPForgeClientEntrypoint.java
@@ -3,9 +3,14 @@ package net.lopymine.ip.entrypoint;
 //? if forge {
 
 /*import net.lopymine.ip.client.InventoryParticlesClient;
+import net.lopymine.ip.client.command.InventoryParticlesCommandManager;
+import net.lopymine.ip.particles.ParticlesConfigsManager;
 import net.lopymine.ip.resourcepack.InventoryParticlesClientReloadListener;
+import net.lopymine.mossylib.loader.MossyLoader;
 import net.lopymine.mossylib.modmenu.*;
 import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.lopymine.ip.modmenu.ModMenuIntegration;
@@ -17,8 +22,11 @@ public class IPForgeClientEntrypoint {
 		ModMenuIntegration integration = new ModMenuIntegration();
 		integration.register(ModLoadingContext.get().getActiveContainer());
 
-		FMLJavaModLoadingContext.get().getModEventBus().<RegisterClientReloadListenersEvent>addListener(event -> {
-			event.registerReloadListener(new InventoryParticlesClientReloadListener());
+		MossyLoader.registerReloadListener(new InventoryParticlesClientReloadListener());
+		MossyLoader.registerCommands(InventoryParticlesCommandManager::register);
+
+		MinecraftForge.EVENT_BUS.<PlayerEvent.PlayerLoggedInEvent>addListener((event) -> {
+			ParticlesConfigsManager.updateCombinedMap();
 		});
 	}