+++ src/main/java/net/lopymine/ip/entrypoint/IPNeoForgeClientEntrypoint.java
@@ -4,19 +4,18 @@ package net.lopymine.ip.entrypoint;
 /*import net.lopymine.ip.InventoryParticles;
 
 import net.lopymine.ip.client.InventoryParticlesClient;
+import net.lopymine.ip.client.command.InventoryParticlesCommandManager;
 import net.lopymine.ip.modmenu.ModMenuIntegration;
-import net.lopymine.ip.resourcepack.InventoryParticlesClientReloadListener;
+import net.lopymine.ip.particles.ParticlesConfigsManager;
+import net.lopymine.ip.resourcepack.*;
+import net.lopymine.mossylib.loader.MossyLoader;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.*;
 import net.neoforged.fml.common.Mod;
-//? if >=1.21.4 {
-import net.neoforged.neoforge.client.event.AddClientReloadListenersEvent;
-//?} else {
-/^
-import net.neoforged.neoforge.client.event.RegisterClientReloadListenersEvent;
-^/
-//?}
+import net.neoforged.neoforge.client.event.*;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.entity.player.PlayerEvent;
 
 @Mod(value = InventoryParticles.MOD_ID, dist = Dist.CLIENT)
 public class IPNeoForgeClientEntrypoint {
@@ -26,15 +25,13 @@ public class IPNeoForgeClientEntrypoint {
 		ModMenuIntegration integration = new ModMenuIntegration();
 		integration.register(container);
 
-		//? if >=1.21.4 {
-		bus.addListener(AddClientReloadListenersEvent.class, event -> {
-			event.addListener(InventoryParticlesClientReloadListener.getId(), new InventoryParticlesClientReloadListener());
+		MossyLoader.registerReloadListener(new InventoryParticlesClientReloadListener());
+		MossyLoader.registerCommands(InventoryParticlesCommandManager::register);
+
+		NeoForge.EVENT_BUS.addListener(PlayerEvent.PlayerLoggedInEvent.class, (event) -> {
+			ParticlesConfigsManager.updateCombinedMap();
 		});
-		//?} else {
-		/^bus.addListener(RegisterClientReloadListenersEvent.class, event -> {
-			event.registerReloadListener(new InventoryParticlesClientReloadListener());
-		});
-		^///?}
+
 	}
 
 }