+++ src/main/java/net/lopymine/ip/tag/ConfigItemTags.java
@@ -1,5 +1,0 @@
-package net.lopymine.ip.tag;
-
-public class ConfigItemTags {
-
-}